pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0", "setuptools"]

[tool.poetry]
name = "autonomy_dev"
version = "0.2.128"
homepage = "https://github.com/8ball030/auto_dev"
description = "A collection of tooling to enable open source development of autonomy tools"
authors = ["8Baller <8ball030@gmail.com>"]
readme = "README.md"
license = "Apache-2.0"
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Natural Language :: English",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
[[tool.poetry.packages]]
include = "auto_dev"

[[tool.poetry.packages]]
include = "tests"
format = "sdist"

[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ["py37", "py38", "py39"]
include = "\\.pyi?$"
exclude = "/(\n    \\.eggs\n  | \\.git\n  | \\.hg\n  | \\.mypy_cache\n  | \\.tox\n  | \\.venv\n  | _build\n  | buck-out\n  | build\n  | dist\n)/\n"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
open-autonomy = "==0.19.3"
open-aea = "==1.64.0"
open-aea-test-autonomy = "==0.19.3"
open-aea-ledger-ethereum = "==1.64.0"
open-aea-ledger-solana = "==1.64.0"
open-aea-ledger-cosmos = "==1.64.0"
open-aea-cli-ipfs = "==1.64.0"
web3 = ">=6.0.0,<7.0.0"
ipfshttpclient = "==0.8.0a2"
typing_extensions = ">=3.10.0.2"
protobuf = ">=3.19.1,<4.25.0"
hypothesis = "==6.122.3"
toml = "==0.10.2"
tomte = {version = "==0.2.17", extras = ["cli", "tests", "ledgers", "black"]}

py-multibase = "==1.0.3"
py-multicodec = "==0.2.1"
py-eth-sig-utils = "*"
grpcio = "==1.69.0"
asn1crypto = "<1.5.0,>=1.4.0"

eth-utils = "==2.3.1"
eth-abi = "==5.1.0"
pycryptodome = "==3.20.0"

rich-click = "^1.8.2"
tbump = "^6.11.0"
requests = "2.28.1"
pytest-xdist = "^3.6.1"
ruff = "^0.5.7"
pydoclint = {extras = ["all"], version = "^0.6.0"}
pydantic = "^2.8.2"
isort = "^5.13.2"
openapi-spec-validator = "0.2.8"
disutils = "^1.4.32.post2"
setuptools = "^75.8.0"


[tool.poetry.group.dev.dependencies]
responses = "^0.23.1"
mkdocs = "^1.6.0"
mkdocs-include-markdown-plugin = "^7.0.0"
mkdocs-mermaid2-plugin = "^1.1.1"
mkdocs-material = "^9.5.47"
mkdocs-material-extensions = "^1.3.1"
mkdocstrings = {extras = ["python"], version = ">=0.25.2,<0.28.0"}
twine = "^5.1.1"
mkdocs-autorefs = "^1.0.1"

[tool.poetry.extras]
all = [
  "pytest",
  "pytest-cov",
  "tbump",
  "pydoclint",
  "ruff",
  "isort",
  "setuptools",
  "pydantic",
  "open-autonomy",
  "open-aea",
  "open-aea-ledger-ethereum",
  "open-aea-ledger-solana",
  "open-aea-ledger-cosmos",
]

[tool.poetry.scripts]
adev = "auto_dev.cli:cli"
[tool.pytest.ini_options]
filterwarnings = [
  "ignore:.*CoverageWarning.*",
  "ignore:datetime.datetime.utcfromtimestamp.*:DeprecationWarning",
  "ignore:There is no current event loop:DeprecationWarning",
  "ignore:DeprecationWarning*.py:DeprecationWarning",
  "ignore:CoverageWarning*.py:",
  "ignore::DeprecationWarning"
]

addopts = "--cov=auto_dev --cov-report html"

